feat: add dedicated ports and connections components
